+++ changelog.md
@@ -8,6 +8,7 @@
 * [Added] Get just outer annotation and find the holes.
 * [Added] Scan segmentation to calculate intra-lobular and extra-lobular surface area
 * [Added] Fixed reading file with no annotation
+* [Added] GLCM texture features used for whole scan segmentation
 
 # 0.15